openai request types
