+++ pkg/openai/edit.go
@@ -5,16 +5,9 @@ import (
 	fm "github.com/redhat-et/copilot-ops/pkg/filemap"
 )
 
-// EditRequest represents user-information which will be passed to OpenAI's edit endpoint.
-type EditRequest struct {
-	Filemap     *fm.Filemap `json:"filemap"`
-	Instruction string      `json:"instruction"`
-}
-
-// EditReply represents the response from OpenAI's edit endpoint.
-type EditReply struct {
-	Filemap *fm.Filemap `json:"filemap"`
-}
+const (
+	EditEndpoint string = "https://api.openai.com/v1/engines/code-davinci-002/edits"
+)
 
 // GetFileEditsFromOpenAI Accepts a Filemap and an instruction string to edit the provided files
 // using a response obtained from OpenAI.
